fix: aplicar correções do code review do UsuarioAdminService
- Extrai PageMetadata e Page<T> para core/models/common.ts; paciente.ts re-exporta via `export type` para manter compatibilidade retroativa
- usuario-admin.ts importa Page de common.ts em vez de paciente.ts
- listarRoles() passa a retornar Observable<UserRole[]> (contrato correto com a API REST); o ROLE_OPTIONS continua para mapeamento de labels no template
- buscarPerfil() reutiliza AuthenticatedUser de auth.ts, eliminando duplicação de interface
- listar() recebe parâmetro sort opcional (padrão 'name') para flexibilidade futura
- Testes: adiciona assertiva de body vazio em ativar/inativar, adiciona sort padrão no teste de page/size customizados, adiciona teste para sort customizado, atualiza listarRoles para UserRole[]
