+++ src/app/core/models/common.ts
@@ -1,0 +1,11 @@
+export interface PageMetadata {
+  totalElements: number;
+  totalPages: number;
+  size: number;
+  number: number;
+}
+
+export interface Page<T> {
+  content: T[];
+  page: PageMetadata;
+}
