+++ src/app/core/models/paciente.ts
@@ -35,14 +35,4 @@ export interface PacienteUpdateDTO {
   endereco?: EnderecoDTO;
 }
 
-export interface PageMetadata {
-  totalElements: number;
-  totalPages: number;
-  size: number;
-  number: number;
-}
-
-export interface Page<T> {
-  content: T[];
-  page: PageMetadata;
-}
+export type { PageMetadata, Page } from './common';
